reorganize planqk plugin for consistency
