+++ components/bpmn-q/modeler-component/extensions/planqk/modeling/PlanQKRenderer.js
@@ -4,12 +4,12 @@ import { is } from "bpmn-js/lib/util/ModelUtil";
 
 import BpmnRenderer from "bpmn-js/lib/draw/BpmnRenderer";
 
-import * as consts from "./utilities/Constants";
+import * as consts from "../utilities/Constants";
 import { getSVG } from "./SVGMap";
 import {
   drawDataElementSVG,
   drawTaskSVG,
-} from "../../editor/util/RenderUtilities";
+} from "../../../editor/util/RenderUtilities";
 
 const HIGH_PRIORITY = 14001,
   TASK_BORDER_RADIUS = 2;
