+++ components/bpmn-q/modeler-component/extensions/planqk/modeling/SVGMap.js
