+++ components/bpmn-q/modeler-component/extensions/planqk/modeling/properties/service-task-properties/SubscriptionProperties.js
